test: Add support for SNIC initiator to test script
To run tests with SNIC initiator the script should be run on the
host. It will do ssh to the SNIC and run initiator. All logs will be
available on the host.

Some other minor improvements and test script cleanup. Now all setup
specific settings are grouped together and may be switched with SETUP
environment variable. New setups may be added to allow parallel
work. To use the setup feature script should be run like this `SETUP=2
./test.sh test4`.

Signed-off-by: Evgeniy Kochetov <evgeniik@nvidia.com>