sock/vma: Load VMA library with dlopen
When shared libraries are used order of constructor calls is not
strictly defined and it is impossible to guarantee correct
initalization order of SPDK and VMA.

This patch loads VMA library with dlopen and all required symbols with
dlsym. LD_PRELOAD of VMA library is not needed anymore.

Path to VMA library is temporary specified with environment variable
SPDK_VMA_PATH. It can later be changed, e.g. to configuration via RPC.

Signed-off-by: Eugene Kochetov <evgeniik@nvidia.com>