nvme: Fix trstring population for custom transport
It is possible to register custom nvme transport with arbitrary name
but it doesn't work because 'spdk_nvme_trid_populate_transport' will
always set trstring to 'CUSTOM' and transport lookup will fail.

With this patch trstring from trid is used for custom transport if it
was set (has non zero length). If trstring is not set it will falback
to default 'CUSTOM' name.

Signed-off-by: Evgeniy Kochetov <evgeniik@nvidia.com>