bdev/nvme: Retry reconnecting ctrlr after seconds if reset failed
Previously reconnect retry was not controlled and was repeated indefinitely.

This patch adds two options, ctrlr_loss_timeout_sec and reconnect_delay_sec,
to nvme_ctrlr and reconnecting_time and reconnect_timer to nvme_ctrlr
to control reconnect retry.

Both of ctrlr_loss_timeout_sec and reconnect_delay_sec are initialized to
zero. This means reconnect is not throttled as we did before this patch.

A few more changes are added.

If ctrlr_loss_timeout_sec and reconnect_delay_sec are 0, ctrlr reset is
repeated very quickly (once per nvme_adminq_poll_period_us) and infinitely.
nvme_adminq_poll_period_us is 10 msec by default. We see nvme_ctrlr is always
resetting if ctrlr reset failed. Not to pend I/O for this case, change
nvme_io_path_is_failed() to return true if reconnect_delay_sec is 0 when
resetting is true or reconnect_start_tsc is not 0.

spdk_nvme_ctrlr_reconnect_poll_async() may continue returning -EAGAIN
infinitely. To check out such exceptional case, use ctrlr_loss_timeout_sec.

Not only ctrlr reset but also non-multipath ctrlr failover is controlled.
So we need to include path failover into ctrlr reconnect.

When the active path is removed and switched to one of the alternative paths,
if ctrlr reconnect is scheduled, connecting to the alternative path is left
to the scheduled reconnect.

If reset or reconnect ctrlr is failed and the retry is scheduled,
switch the active path to one of alternative paths.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: Idec636c4eced39eb47ff4ef6fde72d6fd9fe4f85