blob/esnap: add external snapshot support
An external snapshot can be made from any read-only bdev. A clone of an
external snapshot is known as an external clone. This feature is
implemented by introducing esnap bs devices and extending the extant
support for snapshots and thin provisioning. With thin provisioning, the
back_bs_device is a zeroes device. With external snapshots, the
back_bs_device is an esnap ("external snapshot") device.

This commit introduces what is required to be able create an external
snapshot and perform IO on a blob that uses external snapshots.
Subsequent commits complete the feature set for external snapshots and
provides the same feature for logical volumes.

The feature is more fully described in doc/blob.md.

Contributions-by: Alexey Marchuk <alexeymar@mellanox.com>
Change-Id: I367bacbdcc46f6cddc483036ae8c9bd4fa00e474
Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>