lib/util: add spdk_uuid_clear and spdk_uuid_is_null
Purpose: for use in recognizing the absence of external
snapshots when cloning blobs.

XXX-mg: due to the recent work to add the bdev's UUID as an alias, it is
likely the the consumer of this will transition to using names (or UUID
string alias, whichever is desired).  This would mean that this change
would no longer have a consumer and SHOULD NOT BE MERGED.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: I23d3e881e5584d5b157e1b83bae631329b7154a9