wip-blob/extclone: store seed UUID in internal xattr
Before this, the name of the external snapshot was saved.  That is
fragile because the external name may be subject to whims of external
systems (e.g. nvme namespace order).  Now we use the UUID, which is
consistent across an SPDK-based nvmf target and initiatator.

This also transitions from XATTR (for use by application) to internal
XATTR (for use by blobstore) to find the seed.