wip/blob-extclone: disallow incompatible block sizes
IO operations smaller than the external snapshot's block size would
force the blobstore to use a blobstore-owned buffer to issue external
block sized reads and writes while satisfying smaller requests. While
supporting this would be possible, it would be slow and it's not clear
that there's a need to add this "512e" style emulate at this time.