wip-blob/seed: claim external snapshot bdevs
This registers the seed_if bdev module.  The only function of this
module is to have something to pass to spdk_bdev_module_register_bdev().

An external snapshot may be cloned multiple times, resulting in one
bdev claim.  The claims are done in such a way that the bdev is not
upgraded to read-write.