vbdev_ro: add read-only vbdev
This vbdev is useful when there is a need to be sure that a bdev is not
modified by an application or another vbdev. I ensures that the base
bdev is not written by:

- Opening the base bdev read-only.
- Claiming the base bdev without promoting to read-write. This prevents
  new consumers from opening the base bdev for writing.
- Not implementting write, writev, write_zeroes, discard, nvme_io, or
  any other method that could alter the base bdev.

Multiple ro vbdevs may use the same base bdev. All of the ro vbdevs that
share a base bdev share the claim and IO channels to the base bdev. The
claim and IO channels are released when the last ro vbdev that uses a
base bdev is destroyed.

To do:

- RPC interface
- JSON export (bdev_get_bdevs)
- save_config support.
- Handle surprise remove of base bdev.
- Handle resize of base bdev, maybe.