revert wip-blob/seed: claim external snapshot bdevs
This implementation is problematic because it introduces a new bdev
module under lib/ rather than modules/.  This builds cleanly as a unit
test but the normal full build does not complete.  Rather than hacking
makefiles to get past this, a read-only bdev will be written.

The read-only bdev will have these advantages:

- Will also implement a bdev_submit method that will ensure no data
  mutating operations make it through to the external snapshot.
- Will be usable by other consumers.  As an example, I've wanted a way
  to force blobcli to be read-only so it can be used to examine a
  blobstore that is in use elsewhere.  The read-only bdev could be used
  to ensure that any writes (e.g. recovery due to BS not being marked
  clean) do not happen.