vbdev_ro_ut: poll_threads after create
When a bdev is created, spdk_bdev_register() calls bdev_examine() and
spdk_bdev_wait_for_examine(), the latter of which registers a poller.
The poller calls bdev_wait_for_examine_cb(), which calls
bdev_register_finished().  That is, this poller sends an event that the
bdev was created.  In the course of doing so, it reads bdev->name.  Of
course, the poller is only run once poll_threads() is called.

Unfortunately, poll_thread(), called by poll_threads(), calls
msg_queue_run_batch() prior to calling thread_execute_poller(). Thus,
the bdev's destroy callback is called prior to the poller running to
notify that the bdev was registered. This leads to use-after-free of
the spdk_bdev structure.

TL;DR: This is toxic:

   create_ro_disk(...);
   delete_ro_disk(...);
   poll_threads();

Instead, do this:

   create_ro_disk(...);
   poll_threads();
   delete_ro_disk(...);
   poll_threads();