blob/eio: add bs device that gives EIO for reads
The eio device is intended as a place holder while an external snapshot
is missing. Subsequent commits will allow the eio device to be used when
an external snapshot bdev is not available. Thus, clones of external
snapshots will be able to service IO that involves allocated clusters
while returning EIO for others.

Like the zeroes device, the eio device returns EPERM for writes.