blob/eio: load blob with missing external snapshot
When a blob that is a clone of an external snapshot is loaded, it tries
to open the external snapshot. If it is not available, the eio bs dev is
loaded as back_bs_dev so that IO involving blocks that are not on
previously written clusters return an error. Blocks on previously
written clusters are able to service IO without problems.