bdev_wait: add new bdev module
The wait bdev allows an application or library to await the arrival of
another bdev. It does not support IO.

During creation of a wait bdev, the caller passes a callback. This
callback is invoked when the desired bdev is registered.

This is being added to allow the blobstore to support hot add and remove
of external snapshots, a feature that will be added later in this
series.