bdev: add spdk_bdev_open_by_uuid
In some cases, the bdev uuid is preferred over the name as the primary
identifier for a bdev. This introduces spdk_bdev_open_by_uuid() so that
bdevs can be opened by uuid in a race-free way. Both
spdk_bdev_open_ext() and spdk_bdev_open_by_uuid() share a common
implementation, leading them both to avoid the race that led to the
deprecation and removal of spdk_bdev_open().

XXX-mg this commit has some changes that could go in a separate commit
that adds spdk_bdev_get_by_uuid().