blob: snapshots of external clones
A clone of an external snapshot needs special handling snapshots of this
external clone are created and removed. In particular the following must
exist on the blob that directly references the external snapshot and
must be removed from others:

- SPDK_BLOB_EXTERNAL_SNAPSHOT invalid flag
- BLOB_EXTERNAL_SNAPSHOT_COOKIE internal xattr

See the diagrams and description in blob_extclone_snapshot() in
blob_ut.c for details.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: Ie4125d64d5bac9cfa7d6c7cc9a543d72a169f6ee