nvme_rdma: Call disconnected_qpair_cb when qpair is in disconnected_qpairs list
We want to call disconnected_qpairs_cb only if qpair is actually
disconnected. When we disconnect qpair asynchronously, for qpairs in
the group->disconnected_qpairs list, we want to poll them until
actually disconnected and then call disconnected_qpairs_cb for them.

As a preparation, call disconnected_qpair_cb only for qpairs which is
in the group->disconnected_qpairs list.

For TCP and PCIe transports, disconnecting qpair will continue to be
synchronous for now.

So we change only RDMA transport.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Change-Id: Ifaf6157e1e02fa13f52a66409c9e60fc814d71dd