blob/esnap: reallocate IO channels when needed
When a read of an external snapshot happens on a thread that does not
yet have an IO channel allocated, an async reallocation of IO channels
is triggered. The IO that was submitted is completed with -ENOMEM, which
is an existing clue to callers that the operation may be retried.

Change-Id: I1ee31667f2ada143360a5c79e90dd358d9d6c872
Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>