blob/esnap: perform read after io channel realloc
This commit causes esnap to keep track of the in-flight operation when
IO channel reallocation is needed. Once IO channels are reallocated, the
IO is performed as requested.

In the event that the IO channel is still not available after
reallocation, the original IO is completed with -ENOMEM. Consumers of
the blob may treat -ENOMEM as a transient failure and try again, as has
long been the case.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: I60073bd1e842102559030c3f7e0e5c2ea691912d