setup: default to help
The "config" mode can cause storage that is used in unfamiliar ways to
disappear. It is somewhat common for commands that can be destructive or
disruptive to display a help message when not given explicit instruction
to carry out a task that can be disruptive.

This changes the default mode from "config" to "help".

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: I0f5866693d5cf5b360d00d22281487d7a5ac425d