spdk_top: CPU % should be current interval
The calculation for percent of CPU uses the number of busy cycles in
the current interveral os the numerator and the number of cycles since
the process started as the denominator. This is not useful, as it will
typically show a core or thread that is currently saturated as being
less than 1% busy.

This changes the denominator to use the number of busy + idle cycles in
this interval rather than since the application started.