sock/xlio: Free received packets list on socket close
When disconnect is detected we may still have some unhandled or
partially handled packets in received packets list. We must free them
on socket close. We should have no external references to packets by
that time but just in case we check ref counter and print error in
abnormal case.

Signed-off-by: Evgeniy Kochetov <evgeniik@nvidia.com>