dynamically allocate channels per thread and blob
Each blobstore channel gets a red-black tree indexed by blob ID of
channels for external snapshot bdev channels. This provides lock-free
access to a cache of esnap IO channels.  As esnap IOs are handled,
the esnap IO channels are retrieved from the tree or created and added
if missing.

When a blobstore channel is destroyed, esnap IO channels associated with
that blobstore channel are put (released) and the tree is destroyed.  As
blobs with esnaps are freed, the esnap destroy method releases any
associated esnap IO channels from all of the associated blobstore's
channels.

Change-Id: I3e7033fc514fc916c1ae215538c191e4a1435ab2