blob: external clone creation
This allows for the creation of clones of external snapshots. To keep
the commit size small, the IO path is not yet plumbed. Subsequent
commits will complete the feature set.

See the blob_esnap_create() unit test for examples of how to enable
external snapshots in a blobstore and how to create external clone
blobs.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: I29206b628a2b03b6386a88532565e228df988e0e