blob: prepare sets for external devices
For the various forms for read_bs_dev() and readv_bs_dev() to perform
reads from external devices, the spdk_bs_request_set used for the IO
needs to keep track of the back_bs_dev IO channel as well as the
blobstore's IO channel.

This commit has no change in functionality: it is preparation for a
change in a later commit.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: I8edd9c4bf29bc074194331b42c5ef9d27590ce88