blob: defer bs unload while esnap channels are being destroyed
The following commands lead to the illustrated states.

  bdev_malloc_create -b malloc0
  bdev_lvol_clone_bdev lvs1 malloc0 eclone

     +---------+   +--------+
     | malloc0 |<--| eclone |
     +---------+   +--------+

  bdev_lvol_snapshot lvs1/eclone snap

     +---------+   +------+   +--------+
     | malloc0 |<--| snap |<--| eclone |
     +---------+   +------+   +--------+

  bdev_lvol_clone lvs1/snap eclone

                                +--------+
                              ,-| eclone |
     +---------+   +------+<-'  +--------+
     | malloc0 |<--| snap |
     +---------+   +------+<-.  +-------+
                              `-| clone |
                                +-------+

spdk_blob_unload(snap) is called once for eclone, once for clone, and
once for snap.  The last of these calls happens just before
spdk_bs_unload() is called.  spdk_blob_unload() needs to destroy
channels on each thread.  During this thread iteration, spdk_bs_unload()
starts.  The work performed in the iteration maintains a reference to
the blob, and as such it spdk_bs_unload() cannot do its work until the
iteration is complete.

This makes it so that spdk_bs_unload() recognizes when esnap channel
destruction is happening and causes the blobstore unload to be deferred
until after the channel destruction is complete.

Change-Id: Id9b92ad73341fb3437441146110055c84ee6dc52
Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>