lvol: open external clones with missing snapshots degraded
If an external clone is missing its snapshot the lvol should still open
so that it can be managed and reserve its name and alias in the lvstore
and bdev namespaces.

This allows lvstore to open the external clone and regiesters a
back_bs_dev that will do nothing more than respond to IO requests with
EIO. If an lvol consumer has the lvol open it can perform IO on LBAs
backed by clusters allocated in the blobstore.

TODO: - It seems likely that there will be a desire to prevent opens of lvols
   in this state. If so, consider adding an spdk_bdev_status that
   prevents opens and appropriate functions to set and clear this state.
 - Add tests.
Change-Id: I736194650dfcf1eb78214c8896c31acc7a946b54