vbdev_lvol: degraded open of esnap clones
If an esnap clone is missing its snapshot the lvol should still open so
that it can be managed and reserve its name and alias in the lvstore and
bdev namespaces.

This allows lvstore to open the esnap clone and register a back_bs_dev
that will do nothing more than respond to IO requests with EIO. If an
lvol consumer has the lvol open it can perform IO on LBAs backed by
clusters allocated in the blobstore.

v23.01.nvda merge notes

 - commit f5bd062d03af4da7450cea4b78b923a92c27d254

Change-Id: I736194650dfcf1eb78214c8896c31acc7a946b54
Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>