lvol: introduce lvol_alloc()
There are several places where new lvols are created and each reproduces
much of the same code. External clones will add yet another in lvol.c
and more in unit tests. This introduces lvol_alloc() to minimize the
chance of unintended skew over time.

A side effect of this is that snapshots and clones now inherit clear
method from their parent. Previously they would fall back to the
default. The old behavior seems to be accidental, hence the change.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: Ibf6f79c567e92354ea73e6589c736b1b946731a0