lvol: add support for external snapshots
This provides the lib/lvol wrapper around blobstore's external
snapshots. Later commits make this work with vbdev_lvol.

The blobstore external snapshot implementation stores an opaque cookie
in an internal xattr. Lvstore uses this to store the name or alias of
the bdev that will act as the external snapshot. This is used by the
newly introduced spdk_lvol_create_bdev_clone() to store the bdev name in
the blob's metadata and is used durinb lvol load by the
external_bs_dev_create function, lvs_esnap_dev_create().

Change-Id: I58c7b32b656ad1d21a446e3b91e59e655efac7e4
Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>