vbdev_lvol: load external snapshot via examine callback
With this commit when an external snapshot device becomes available, it
is hot-plugged into the lvol and the lvol is no longer degraded.

The arrival of the bdev is noticed by vbdev_lvol's examine_disk()
callback, vbdev_lvol_examine(). Each lvstore's missing external
snapshots tree is searched for lvols matching the new bdev's name and
uuid string.  Each match triggers the creation of a struct spdk_bs_dev
which is registered with the appropriate blob. Any number of blobs in
any number of blobstores may concurrently use the same bdev as an
external snapshot. Each blob that uses a bdev will have its own
bdev_desc.

Note that the examine callbacks are called before any bdev aliases can
be registered so aliases are not viable as external snapshot names.

Change-Id: I0e2eb26375c62043b0f895197b24d6e056905aa2
Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>