include/mutex.h: safer mutexes
The default mutex initialized by pthread_mutex_init() is not as useful
as it could be. If accidentally used with recursion, it leads to
recursion. There is also no way to programmatically verify that a lock
is held or not held. This introduces a few inline functions to easily
use errorcheck mutexes to catch locking problems before they turn into
worse problems.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: I6dd6493ab5f5532ae69e20654546405a507eb594