bdev: hold mutex while changing bdevs list
bdev_unregister_unsafe() documents that g_bdev_mgr.mutex must be held.
It seems the reason for this would be due to it potentially removing
a bdev from g_bdev_mgr.bdevs. If it is important to hold this mutex
while removing an item from the list here, it's probably important that
another thread isn't simultaenously adding or removing items elsewhere.

This commit ensures that g_bdev_mgr.mutex in all places where
g_bdev_mgr.bdevs is altered. There remain many places where the list is
accessed without holding the lock. This doesn't seem safe to me, but
I've not studied the TAILQ macros closely.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: I7dcd326937f296fd956e533941fb878d247c72d8