thread: DEBUG spdk_mutex should record line number
With a DEBUG build, each call to spdk_mutex_init(),
spdk_mutex_destroy(), spdk_mutex_lock(), and spdk_mutex_unlock() will
store the line number of the caller in smutex->line. When an abort()
happens due to errorcheck, it is now trivial to find the caller that
left the SPDK mutex in an unexpected state.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: I7ff38f2a31db03e19d0cfb26c26e62bdec7265f8