bdev: wrong lock taken in spdk_bdev_unregister
While spdk_bdev_unregister() is checking to see if the bdev is already
unregistering or removing, it erroneously takes g_bdev_mgr.lock.  If any
lock is taken, it should be bdev->internal.mutex.

The right fix would seem to be to bdev->internal.mutex a few lines
earlier so that the bdev->internal.status check is performed with the
lock held. When this is done, the error checking on spdk_mutex triggers
aborts in a few places in bdev_ut.c when free_bdev() is called. This is
because spdk_bdev_unregister() is called twice: once explicitly in
the test and again by free_bdev(). By the time that it is called the
second time, bdev_unregister() has already destroyed
bdev->internal.mutex and a typical bdev module would have freed the
bdev. The test is special in that its destruct callback is a no-op,
leading to this oddball case of using a lock after destruction but not a
use-after-free bug.

During this check, holding g_bdev_mgr.mutex is of no benefit, holding
bdev->internal.mutex seems to not be required, and holding
bdev->internal.mutex will cause problems for unit tests. Thus, the check
of bdev->internal.status for an unregistering or removed status is now
done with no lock held.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: I0645f73711853ae7cfa494266df2641ddb18461e