bdev: call examine_disk() for all claim holders
If multiple claims exist on a bdev, examine_disk() is called for each of
them.

Traditionally, examine_disk() has been called once per module. So long
as there is only one descriptor with a claim per module, this is still
the case. If there are multiple descriptors with claims from the same
module, the module will have its examine_disk() callback invoked
multiple times.

There is no clear reason that the multiple calls with no unique
parameter is desirable, but it's also not clear that this is a problem
that needs to be solved now. If a bdev module wishes to receive only one
notification per bdev, it should not make claims using multiple
descriptors.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: I0a6dc3e4bd1da20bbcbddf97a16e04c62c82354c