test: new test API to look for specific messages
This commit adds a way for unit tests to ensure that messages that
should be logged are logged and unexpected messages are not logged.

While the first intended use of this is to verify that deprecation
messages are printed, it can (and perhaps should) be used in all unit
tests to catch unexpected messages that may be indicative of problems
not caught in typical error checking. Use of this API also cleans up
unit test output, making it much easier to spot anomalies.

Change-Id: Iba0e1d4224e357c7941e5ff0c618535fdd60215c
Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>