DO NOT MERGE: bdev: test for bdev internal lock on wrong thread
With bdev management operations all on the app thread, it should be
possible to do away with bdev->internal.lock. This commit adds a
deprecation and updates existing "wrong thread" tests to support this
theory.

Note: I'm not sure that there is benefit in this largely duplicate
deprecation being merged. I generated it primirialy to determine whether
there were any bdev API functions that missed the previous deprecation.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: I3a2a3cf529000ccf1708540cb39642d556ce0aeb