fio_plugin: bdev management on app thread
With the bdev_mgmt_wrong_thread deprecation the fio plugin needs to open
and close bdevs on the app thread.

The bdev opens and closes happen in synchronous calls from fio. The
spdk_dev_open_ext() calls are passed to the app thread then the calling
thread waits on a condition variable for spdk_bdev_open_ext() to return.
The spdk_bdev_close() calls have no return value and no critical timing
aspects, so they are handled as "fire and forget.'

Change-Id: I6719723e6a9f7aaa217b1ea5879dd011f14039be
Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>