bdevperf: move bdev management to the app thread
With the bdev_mgmt_wrong_thread deprecation, bdev management operations
need to move to the SPDK app thread. This commit moves bdev open and
close calls to the app thread.

A side effect of opening the bdev on the app thread is that
notifications via the event callback now happen on the app thread. The
SPDK_BDEV_EVENT_REMOVE event leads to a poller removal that needs to be
sent to the appropriate job thread.

Change-Id: Ib6f90e327a00e0e0bd823b52c922995ad83c4408
Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>