fio_plugin: bdev calls on spdk thread
This is a rework of commit 7722996dd247fbe7566f578a06526e32b7fc4ebc.

As the locks in bdev.c transition to spdk_spinlock, callers of
spdk_bdev_get_by_name() and any other function that uses an SPDK
spinlock needs to be executing from an spdk_thread.  This commit reverts
the earlier commit then fixes it with a different approach than was
previously used.

This commit splits spdk_fio_setup() into two parts:

1. SPDK library initialization: this part may run before SPDK is
   initialized and as such will run once when no SPDK thread is
   available. After performing any required SPDK library initialization,
   it sends a message to the app thread to run setup_on_app_thread().
2. fio initialization in setup_on_app_thread(): this part iterates
   through all of the files in each fio thread's configuration and
   provides initialization on each of them.

The fio plugin API only supports synchronous calls. Thus,
spdk_fio_setup() waits for the completion of setup_on_app_thread() using
a condition variable. The code is arranged to make it easy to use this
same mechanism for a future commit that will be needed with the
bdev_mgmt_wrong_thread deprecation.

Fixes issue #2818

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>
Change-Id: I5c82542bb9e9e6b8823e9d1da293050bba336b65